+++ src/main/java/roomescape/admin/reservation/RequestReservation.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.admin.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
